requirements.txt
pandas>=2.0
requests>=2.28
openpyxl>=3.1
flask>=3.0
flask-cors>=4.0
python-dotenv>=1.0
